Refactor name propery of the transfer
Name property belongs logically only on Transfer entity and not on the
client or base API request class.
This commit adds the property on the Transfer class and from there we
pass it around and use it to call the CreateTransfer API requests
objects. This way only this API request class has name as property
instead of all of them.